+++ CHANGELOG.md
@@ -6,11 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
-### Changed
+### Added
 
 -   Add typings.
--   **Breaking:** The list of props is now passed to `makeWebComponents` as the second argument instead of being a property on the function.
 -   Support props without attribute, which can be any type.
+-   Add prop setter functions. These functions don't trigger a render, but they can dispatch events.
+
+### Changed
+
+-   **Breaking:** The list of props is now passed to `makeWebComponents` as the second argument instead of being a property on the function.
 -   **Breaking:** `props` are not mirrored to attributes anymore. Use `attrs` for this usecase, but note that these can only hold strings.
 
 ## [0.0.4]